feat: display "runs-on" only when it is set
By default, GH actions constructs the check name based on matrix input values. It does skip the `null`, so if we have a single, default `runs-on` (i.e. it is left undefined in the input), we may as well skip it and set the default only in the workflow itself.